00029 - In-placemultiline editor added
