+++ src/app/store.ts
@@ -5,6 +5,7 @@ import dealTagsReducer from "../features/dealTags/dealTagsSlice";
 import eventsReducer from "../features/events/eventsSlice";
 import eventNotesReducer from "../features/eventNotes/eventNotesSlice";
 import enumsReducer from "../features/enums/enumsSlice";
+import textareaReducer from "../features/textarea/textareaSlice";
 
 export const store = configureStore({
   reducer: {
@@ -14,6 +15,7 @@ export const store = configureStore({
     events: eventsReducer,
     eventNotes: eventNotesReducer,
     enums: enumsReducer,
+    textarea: textareaReducer,
   },
 });
